Web app build v01/Try1 Flask/FlaskWebProject1/FlaskWebProject1/requirements.txt
Flask~=1.1
mysqlclient
mysql-connector-python
pymysql
cx_Oracle
flask_mysqldb
flask-sqlalchemy
flask-migrate
flask-session
redis
flask_redis
python_dotenv
flask_wtf
wtforms
